==> code/4-ptk-uniformity/2-ordinal-relationships.Rmd <==
---
title: "2. /ptk/ uniformity - ordinal relationships"
author: "Khia A Johnson"
date: "4/1/2021"
output: html_document
---

## setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggthemes)

theme_set(
  theme_clean() + 
  theme(
      legend.title = element_blank(),
      legend.text = element_text(size=8),
      legend.position = 'bottom',
      axis.title.x = element_text(size = 8, face='bold'),
      axis.title.y = element_text(size = 8, face='bold'),
      axis.text.x = element_text(size = 8),
      axis.text.y = element_text(size = 8),
      strip.text = element_text(size = 8, face='bold'),
      )
  )
```

```{r}
df <- read_csv('dfs/ptk_clean_20210602.csv') # this version used for interspeech revisions
head(df)
```

## Oridnal relationship proportions
Proportion of talkers adhering to expected ptk duration relationships (more posterior --> longer) given prior work on connected (but not spontaneous) speech. These numbers are all over the map, though if we take the estimates for /p/ with a grain of salt (some individuals have very few tokens), the t-k comparison more or less follows expectations.
```{r}
df %>%
  group_by(talker, segment, language) %>%
  summarise(mv = mean(vot_ms)) %>%
  ungroup() %>%
  pivot_wider(names_from = 'segment', values_from = 'mv') %>%
  ungroup() %>%
  drop_na() %>%
  group_by(language) %>%
  summarise(
    pt = sum(p<t)/n(), 
    tk = sum(t<k)/n(), 
    pk = sum(p<k)/n(),
    ptk = sum(p<t & t<k)/n(),
    n=n()
    )
```

```{r}
df %>%
  group_by(talker, segment, language) %>%
  summarise(mv = mean(vot_ms)) %>%
  ungroup() %>%
  pivot_wider(names_from = 'segment', values_from = 'mv') %>%
  ungroup() %>%
  mutate(
    order = if_else(p<t & t<k, 'ptk', ''),
    order = if_else(p<k & k<t, 'pkt', order),
    order = if_else(t<p & p<k, 'tpk', order),
    order = if_else(t<k & k<p, 'tkp', order),
    order = if_else(k<p & p<t, 'kpt', order),
    order = if_else(k<t & t<p, 'ktp', order),
  ) %>%
  select(talker, language, order) %>%
  pivot_wider(names_from=language, values_from=order) %>%
  mutate(same_order = Cantonese==English) %>%
  drop_na() %>%
  summarize(sum(same_order))
```

But... looking at just the talkers who produce more /p/ tokens doesn't really change those proportions in ways that better reflect expected ordinal relationships. Takeaway? These might be differences between spontaneous and read speech. 
```{r}
df %>%
  group_by(talker, segment, language) %>%
  summarise(mv = mean(vot_ms), n=n()) %>%
  filter(n>=10) %>%
  select(talker, segment, language, mv) %>%
  ungroup() %>%
  pivot_wider(names_from = 'segment', values_from = 'mv') %>%
  ungroup() %>%
  drop_na() %>%
  group_by(language) %>%
  summarise(
    pt = sum(p<t)/n(), 
    tk = sum(t<k)/n(), 
    pk = sum(p<k)/n(),
    n=n()
    )
```

Range of counts that the means are based on -- far fewer per talker for /p/
```{r}
df %>%
  group_by(talker, segment, language) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  group_by(language, segment) %>%
  summarise(min(n), mean(n), max(n))
```

## Visualizing ordinal relationships



## Visualizing within talker segment relationships
```{r fig.width=7, fig.height=3.5}
std.error <- function(x, na.rm = T) {
  sqrt(var(x, na.rm = na.rm)/length(x[complete.cases(x)]))
}

df$segment <- as_factor(df$segment)
df$segment <- forcats::fct_relevel(df$segment, 'p','t','k')
levels(df$segment)
```

Ordinal relationships within talker by language, in graphical form -- probably wont include this in the paper
```{r fig.width=5.5, fig.height=6.5}
df %>%
  filter(str_detect(talker, 'VM')) %>%
  group_by(talker, language, segment) %>%
  summarise(mean_vot = mean(vot_ms), se_vot = std.error(vot_ms), n=n()) %>%
  ungroup() %>%
  mutate(lower = mean_vot - se_vot, upper = mean_vot + se_vot) %>%
  mutate(segment = str_to_upper(segment), language = if_else(language=='English', 'E', 'C')) %>%
  ggplot(aes(y=mean_vot, x=language, color=segment, shape=segment, label=segment)) +
    geom_errorbar(width=0.85, size=0.5 ,aes(ymin = lower, ymax = upper), position=position_dodge(width = 0.75), alpha=0.35) +
    scale_color_viridis_d(option = 'magma', begin = 0, end = 0.7, breaks = c('P','T','K')) +
    #scale_shape_discrete(breaks = c('P','T','K')) +
    geom_text(size = 2.5 , position=position_dodge(width = 0.750), fontface = "bold") +
  ylab('Mean VOT (ms)') +
  xlab('') +
  coord_flip() +
  facet_wrap(~talker, ncol=3) +
    theme(
      panel.grid.major.x = element_line(color = 'gray', linetype='dotted'),
      legend.position = 'none'
    )

ggsave('../../text/figures/ch4_ordrel_vm_5in.png', height = 6.5, width = 5, units='in')
```



```{r}
df %>%
  ggplot(aes(x=vot_ms, color=segment, linetype=language)) +
  geom_boxplot()
```




```{r fig.height=4, fig.width=7.5}
df %>%
  group_by(talker, language, segment) %>%
  summarise(mean_vot = mean(vot_ms), se_vot = std.error(vot_ms), n=n()) %>%
  ungroup() %>%
  mutate(language.dodged = if_else(language=='Cantonese', 1, 0),
         language.dodged = if_else(segment=='p', language.dodged+0.2, language.dodged),
         language.dodged = if_else(segment=='k', language.dodged-0.2, language.dodged)
         ) %>%
  ggplot(aes(y=language.dodged, x=mean_vot, shape=segment, color=segment, group=segment)) +
    geom_point(size=1) +
    geom_errorbar(alpha=0.85, size=.5, width = .2, aes(xmin = mean_vot-se_vot, xmax = mean_vot+se_vot)) +
    geom_line(alpha=0.5, size=.5) +
    scale_color_viridis_d(option = 'inferno', begin = 0.1, end = 0.75) +
    scale_y_continuous(breaks = c(0,1), labels = c('E', 'C')) +
    facet_wrap(~talker, ncol=7)+
    labs(y="",  x="VOT (ms)") +
    theme(
      strip.text = element_text(size=6),
      legend.position = 'bottom',
      legend.spacing.y = unit(0, 'cm')
    )
# ggsave('figs/ptk_vot_individuals.png', height = 4.5, width = 7.5, units='in')
# ggsave('figs/ptk_vot_individuals_vert.png', height = 8.5, width = 3.5, units='in') # Used in interspeech paper with manually moved legend - don't re-run; different sizes for paper and poster
```
