==> code/4-ptk-uniformity/4-mixed-effects-model.Rmd <==
---
title: "2. /ptk/ uniformity - ordinal relationships"
author: "Khia A Johnson"
date: "4/1/2021"
output: html_document
---
## Imports
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(lme4)
library(lmerTest)
library(optimx)
library(wec)
library(emmeans)
library(brms)
library(ggthemes)
library(tidybayes)
library(bayesplot)

options(scipen=999)

theme_set(
  theme_clean() + 
  theme(
      legend.title = element_blank(),
      legend.text = element_text(size=8),
      legend.position = 'bottom',
      axis.title.x = element_text(size = 8, face='bold'),
      axis.title.y = element_text(size = 8, face='bold'),
      axis.text.x = element_text(size = 8),
      axis.text.y = element_text(size = 8),
      strip.text = element_text(size = 8, face='bold'),
      )
  )
```

```{r}
df <- read_csv('dfs/ptk_clean_20210602.csv')
df
```

Center VOT and standardize average phone duration
```{r}
df$vot_z <- scale(df$vot_ms, center=TRUE, scale=TRUE)
df$vot_scaled <- scale(df$vot_ms, center=FALSE, scale=TRUE)

df$apd_z <- scale(df$avg_phone_dur_ms, center=TRUE, scale=TRUE)
```

Set up and inspect weighted effect coding
```{r}
df$is_post_pausal <- as_factor(df$is_post_pausal)
contrasts(df$is_post_pausal) <- contr.wec(x=df$is_post_pausal, omitted = FALSE)
contrasts(df$is_post_pausal)

df$segment <- as_factor(df$segment)
contrasts(df$segment) <- contr.wec(x=df$segment, omitted = 'p')
contrasts(df$segment)

df$language <- as_factor(df$language)
contrasts(df$language) <-  contr.wec(x=df$language, omitted = 'Cantonese')
contrasts(df$language)
```

```{r fig.width=3}
df %>%
  ggplot(aes(x=vot_scaled)) +
  geom_histogram()

df %>%
  ggplot(aes(x=vot_ms)) +
  geom_histogram()
```

```{r}
get_prior(
  vot_z ~ 1 + segment * language + apd_z + is_post_pausal + (1 + segment * language | talker) + (1 | word),
    family = gaussian(),
    data = df
)
```


Same model structure as Interspeech, but bayesian
```{r}
m <- 
  brm(
  vot_z ~ 1 + segment * language + apd_z + is_post_pausal + (1 + segment * language | talker) + (1 | word),
    family = gaussian(),
    data = df,
    prior = c(
      prior(student_t(3, 0, 2.5), class = Intercept),
      prior(normal(0, 1), class = b),
      prior(lkj(2), class = cor) 
      ),
    chains = 4, cores = 4, iter = 5000, warmup = 2500,
    file = 'fits/m.gaussian.final'
  )
```

```{r}
betas <- m %>%
  get_variables() %>%
  head(8) %>%
  array()
```

Model checks
```{r fig.width=12, fig.height=4}
mcmc_trace(m, pars = betas, facet_args = c(ncol = 4))
```

```{r}
mcmc_plot(m, type='rhat_hist')
```

```{r}
mcmc_plot(m, type='neff_hist')
```

Get variable names for use in later plotting
```{r}
as_tibble(get_variables(m))
```

```{r}
mean(df$vot_ms) / sd(df$vot_ms)
```

```{r}
round(fixef(m),2)
```


```{r}
summary(m)

```


```{r}
popsamples <- posterior_samples(m) %>% 
  select(
    Intercept = b_Intercept,
    `Place (T)` = b_segmentt,
    `Place (K)` = b_segmentk,
    `Language (English)` = b_languageEnglish,
    `Place (T) × Language (English)` = `b_segmentt:languageEnglish`,
    `Place (K) × Language (English)` = `b_segmentk:languageEnglish`,
    `Average Phone Duration` = b_apd_z,
    `Preceding Pause (True)` = b_is_post_pausalTRUE,
    ) %>%
  pivot_longer(everything(), names_to = 'Parameter') %>%
  mutate(Parameter = as_factor(Parameter))


```


```{r}
popsamples %>%
  mutate(outside_rope = value < -0.1 | value > 0.1) %>%
  arrange(value, desc=T) %>%
  group_by(Parameter) %>%
  summarise(mean(value), sum(outside_rope)/100)
```

```{r}
popsamples %>%
  filter(Parameter != 'Intercept') %>%
  mutate(value = value ) %>%
  ggplot(aes(x = value, y = Parameter)) +
  stat_halfeye(.width = .95, size = 1/2) +
  scale_y_discrete(limits = rev) +
  annotate("rect", xmin=-0.1, xmax=0.1, ymin=0, ymax=Inf, fill = '#FEB77EFF' , alpha=0.5)

ggsave('../../text/figures/ch4_poppar_5in.png', height = 3, width = 5, units='in')
```


```{r }
sdsamples <- posterior_samples(m) %>%
  select(
    `Intercept: Word` = sd_word__Intercept,
    `Intercept: Talker` = sd_talker__Intercept,
    `By-talker slope: Place (T)` = sd_talker__segmentt,
    `By-talker slope: Place (K)` = sd_talker__segmentk,
    `By-talker slope: Language (English)` = sd_talker__languageEnglish,
    `By-talker slope: Place (T) × Language (English)` = `sd_talker__segmentt:languageEnglish`,
    `By-talker slope: Place (K) × Language (English)` = `sd_talker__segmentk:languageEnglish`
    ) %>%
  pivot_longer(everything(), names_to = 'Parameter', values_to = 'SD') %>%
  mutate(Parameter = as_factor(Parameter))
```


```{r}
sdsamples %>%
  ggplot(aes(x = SD, y = Parameter)) +
  stat_halfeye(.width = .95, size = 1/2) +
  scale_y_discrete(limits = rev) +
  xlab('Standard Deviation')

ggsave('../../text/figures/ch4_grppar_5in.png', height = 3, width = 5, units='in')
```

```{r fig.height=10, fig.width=12}
rsamples <- posterior_samples(m) %>% 
  select(starts_with("r_talker")) %>% 
  pivot_longer(everything()) %>% 
  separate(name, into = c('talker', 'parameter'), sep = ',') %>%
  mutate(talker = str_replace(talker, 'r_talker\\[', ''),
         parameter = str_replace(parameter, '\\]',''),
         parameter = str_replace(parameter, 'languageEnglish', 'Language (English)'),
         parameter = str_replace(parameter, 'segmentk', 'Place (K)'),
         parameter = str_replace(parameter, 'segmentt', 'Place (T)'),
         parameter = str_replace(parameter, ':', ' ×\n'),
         parameter = as_factor(parameter)
         )

levels(rsamples$parameter) <- c("Intercept", "Place (T)", "Place (K)", "Language (English)", "Place (T) ×\nLanguage (English)", "Place (K) ×\nLanguage (English)")
```

```{r fig.width=5, fig.height=7}
rsamples %>%
  ggplot(aes(x = value, y = talker, ))+#color = parameter)) +
  stat_pointinterval(shape=18) +
  geom_vline(xintercept = 0) +
  scale_y_discrete(limits = rev) +
  annotate("rect", xmin=-0.1, xmax=0.1, ymin=0, ymax=Inf, fill = '#FEB77EFF' , alpha=0.5) +
  facet_wrap(~parameter, ncol=3) +
  ylab('Talker') + 
  xlab('Estimate')

ggsave('../../text/figures/ch4_grpparvar_5in.png', height = 8.05, width = 5.75, units='in')

```

```{r}
posterior_samples(m) %>% 
  select(starts_with("r_word")) %>% 
  pivot_longer(everything()) %>%
  group_by(name) %>%
  summarise(mv=mean(value)) %>%
  arrange(desc(mv))
```

This makes for a ~5 ms interval that tracks with the gradient effects in McMurray et al (2002)
```{r}
sd(df$vot_ms) * 0.1
```

```{r}
levels(df$segment) <- c('p', 't', 'k')

df %>%
  ggplot(aes(y=vot_ms, x=language)) +
  geom_violin(alpha=0.5, aes(fill=language)) +
  geom_line(aes(group = talker), alpha=0.5)+
  
  scale_fill_colorblind() + 
  facet_wrap(~segment)
```

```{r}
m$formula
```

Posterior predictions for average talker/word
```{r}
newdata <- data.frame(
  language = factor(c("Cantonese", "English","Cantonese", "English", "Cantonese", "English")),
  segment = factor(c("p", "p", "t", "t", "k", "k")),
  apd_z = c(0,0,0,0,0,0),
  is_post_pausal = c(F,F,F,F,F,F)
)

pp <-predict(m, newdata = newdata, re_formula = ~ segment * language + apd_z + is_post_pausal, summary = FALSE)
```


```{r}
sxl <- conditional_effects(m, effects = 'segment:language', method = 'fitted', prob=0.95)

plot(sxl, plot = F)[[1]] + 
  scale_color_viridis_d(option='magma', end=0.8) +
  scale_x_discrete(limits=rev, breaks = c('p', 't', 'k'), labels = c('P', 'T', 'K')) +
  xlab('Place') +
  ylab('Estimated Mean VOT (z)')

ggsave('../../text/figures/ch4_conditionaleffects_4in.png', height = 3, width = 4, units='in')

```


```{r}
as_tibble(pp) %>%
  select(Cantonese_P = V1, English_P = V2, Cantonese_T = V3, English_T = V4, Cantonese_K = V5, English_K = V6) %>%
  pivot_longer(everything()) %>%
  separate(name, into = c('Language', 'Phone')) %>%
  mutate(Phone = as_factor(Phone)) %>%
  ggplot(aes(x=value, color=Language)) +
  geom_density()+
  # scale_x_continuous(limits = c(0,3))+
  geom_vline(xintercept=0, color='gray', linetype='dotted') +
  facet_wrap(~Phone) +
  scale_color_viridis_d(option='magma', end=0.8) +
  ylab('Density') +
  xlab('Predicted value')
```

```{r}
as_tibble(pp) %>%
  select(Cantonese_P = V1, English_P = V2, Cantonese_T = V3, English_T = V4, Cantonese_K = V5, English_K = V6) %>%
  pivot_longer(everything()) %>%
  separate(name, into = c('Language', 'Phone')) %>%
  mutate(Phone = as_factor(Phone)) %>%
  ggplot(aes(x=value, color=Phone)) +
  geom_density()+
  # scale_x_continuous(limits = c(0,3))+
  geom_vline(xintercept=0, color='gray', linetype='dotted') +
  facet_wrap(~Language) +
  scale_color_viridis_d(option='magma', end=0.8) +
  ylab('Density') +
  xlab('Predicted value')
```


## Revisiting model structure


no interaction
```{r}
m.noint <- 
  brm(
  vot_z ~ 1 + segment + language + apd_z + is_post_pausal + (1 + segment + language | talker) + (1 | word),
    family = gaussian(),
    data = df,
    prior = c(
      prior(student_t(3, 0, 2.5), class = Intercept),
      prior(normal(0, 1), class = b),
      prior(lkj(2), class = cor) 
      ),
    chains = 4, cores = 4, iter = 5000, warmup = 2500,
    file = 'fits/m.gaussian.noint'
  )
```

```{r}
m.compare <- add_criterion(m, c("loo", "waic"))
m.noint.compare <- add_criterion(m.noint, c("loo", "waic"))

loo_compare(m.compare, m.noint.compare)
```


```{r}
model_weights(m.compare, m.noint.compare, weights = 'loo')
```

ok so model comparison says that the interaction is doing work


## How does exclusion of word pooling change things


no word pooling
```{r}
m.noword <- 
  brm(
  vot_z ~ 1 + segment * language + apd_z + is_post_pausal + (1 + segment * language | talker),
    family = gaussian(),
    data = df,
    prior = c(
      prior(student_t(3, 0, 2.5), class = Intercept),
      prior(normal(0, 1), class = b),
      prior(lkj(2), class = cor) 
      ),
    chains = 4, cores = 4, iter = 5000, warmup = 2500,
    file = 'fits/m.gaussian.noword'
  )
```


```{r}
summary(m.noword)
```


```{r}
conditional_effects(m.noword, effects = 'segment:language', method = 'fitted', prob=0.95)
```



```{r}
popsamples.noword <- posterior_samples(m.noword) %>% 
  select(
    Intercept = b_Intercept,
    `Place (T)` = b_segmentt,
    `Place (K)` = b_segmentk,
    `Language (English)` = b_languageEnglish,
    `Place (T) × Language (English)` = `b_segmentt:languageEnglish`,
    `Place (K) × Language (English)` = `b_segmentk:languageEnglish`,
    `Average Phone Duration` = b_apd_z,
    `Preceding Pause (True)` = b_is_post_pausalTRUE,
    ) %>%
  pivot_longer(everything(), names_to = 'Parameter') %>%
  mutate(Parameter = as_factor(Parameter))

popsamples.noword %>%
  filter(Parameter != 'Intercept') %>%
  mutate(value = value ) %>%
  ggplot(aes(x = value, y = Parameter)) +
  stat_halfeye(.width = .95, size = 1/2) +
  scale_y_discrete(limits = rev) +
  annotate("rect", xmin=-0.1, xmax=0.1, ymin=0, ymax=Inf, fill = '#FEB77EFF' , alpha=0.5)
```

```{r}
sdsamples.noword <- posterior_samples(m.noword) %>%
  select(
    `Intercept: Talker` = sd_talker__Intercept,
    `By-talker slope: Place (T)` = sd_talker__segmentt,
    `By-talker slope: Place (K)` = sd_talker__segmentk,
    `By-talker slope: Language (English)` = sd_talker__languageEnglish,
    `By-talker slope: Place (T) × Language (English)` = `sd_talker__segmentt:languageEnglish`,
    `By-talker slope: Place (K) × Language (English)` = `sd_talker__segmentk:languageEnglish`
    ) %>%
  pivot_longer(everything(), names_to = 'Parameter', values_to = 'SD') %>%
  mutate(Parameter = as_factor(Parameter))

sdsamples.noword %>%
  ggplot(aes(x = SD, y = Parameter)) +
  stat_halfeye(.width = .95, size = 1/2) +
  scale_y_discrete(limits = rev) +
  xlab('Standard Deviation')
```


```{r}
sdsamples.noword$model <- 'no-word-ints'
```

```{r}
sdsamples %>%
  mutate(model = 'full') %>%
  bind_rows(sdsamples.noword) %>%
  ggplot(aes(x = SD, y = Parameter, fill=model)) +
  stat_halfeye(.width = .95, size = 1/2, alpha=0.75) +
  scale_y_discrete(limits = rev) +
  xlab('Standard Deviation') 
```



sd
